Devlope_and_help_document/requirements.txt
pandas
python-docx
pyQt5
pyqt5-tools
xlrd
Pillow
pymysql
lxml
asciitomathml